requirements.txt
ray==0.8.0
sanic==19.9.0
numpy==1.17.3
tensorflow==2.0.0
torch==1.3.1
gym==0.15.4
pulsar-client==2.4.0
protobuf==3.11.0
apache-bookkeeper-client==4.10.0


